# Improve performance of findings retrieval